Update 5_dada2.R
Jeff/Justin's code set the number of samples to learn error rates to 36, but this would mean the code breaks if there's less than 36 samples. I've adjusted the code to calculate how many samples are needed to hit ~1 million reads (which is the target) and also added a check where, if there's not enough reads to hit 1 million, then a warning message is printed but all reads available are used.